# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Code Master API",
        "description": "Adaptive Learning Platform - Multi-tenant SaaS",
        "contact": {
            "email": "dev@codemaster.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api/v1",
            "description": "Local Sandbox"
        },
        {
            "url": "https://api.codemaster.com/api/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/tracks": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "List all tracks - public, paginated, cached 1h",
                "operationId": "ad1baacaa4dfb415363918e8e58e0833",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated tracks list"
                    }
                }
            }
        },
        "/tracks/{id}": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Show track details with courses",
                "operationId": "8fc9f230b412ad82102ca0395758fc19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track with nested courses"
                    },
                    "404": {
                        "description": "Track not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tracks/{id}/enroll": {
            "post": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Enroll in a track",
                "operationId": "a7183eb676f708550d3f561935b02f86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrolled successfully"
                    },
                    "400": {
                        "description": "Already enrolled or threshold not met"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tracks/my-tracks": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get enrolled tracks for authenticated user",
                "operationId": "6c59d95436c4c8222af025fc8622e10a",
                "responses": {
                    "200": {
                        "description": "Enrolled tracks with progress"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}